+++ tassaron_flask/__main__.py
@@ -1,6 +1,6 @@
 """
 Entrypoint for `python -m modulename` starts a Flask development server
 """
-from .run import application
+from tassaron_flask.run import application
 
 application.run()
