+++ tassaron_flask/controllers/main/routes.py
@@ -18,7 +18,7 @@ main_routes = Blueprint("main", __name__)
 
 
 # import images at this point to ensure that all of main_routes is defined
-from .images import *
+from tassaron_flask.controllers.main.images import *
 
 
 def generic_url_for(rule):
