+++ tassaron_flask/controllers/shop/cart_api.py
@@ -4,7 +4,7 @@ Each endpoint returns a response of success or not, to update the client-side re
 """
 from flask import Blueprint, request, session
 from tassaron_flask.helpers.main.plugins import db
-from ...models.shop.inventory_models import Product
+from tassaron_flask.models.shop.inventory_models import Product
 
 
 blueprint = Blueprint(
