+++ tassaron_flask/controllers/shop/inventory.py
@@ -3,8 +3,8 @@ from flask_login import current_user
 from werkzeug.datastructures import MultiDict
 from tassaron_flask.blueprint import Blueprint
 from tassaron_flask.helpers.main.plugins import db
-from ...forms.shop.inventory_forms import ProductForm
-from ...models.shop.inventory_models import Product
+from tassaron_flask.forms.shop.inventory_forms import ProductForm
+from tassaron_flask.models.shop.inventory_models import Product
 import os
