+++ tassaron_flask/models/main/models.py
@@ -1,4 +1,4 @@
-from ...helpers.main.plugins import plugins
+from tassaron_flask.helpers.main.plugins import plugins
 from itsdangerous import TimedJSONWebSignatureSerializer
 from datetime import datetime
 from flask import current_app
