+++ tests/test_email_verification.py
@@ -1,4 +1,4 @@
-from tassaron_flask.models.main import create_app, init_app
+from tassaron_flask.helpers.main.app_factory import create_app, init_app
 from tassaron_flask.helpers.main.plugins import db
 from tassaron_flask.models.main.models import User
 from tassaron_flask.helpers.main.email import *
