+++ tests/test_path_traversal.py
@@ -1,4 +1,4 @@
-from tassaron_flask.models.main import create_app, init_app
+from tassaron_flask.helpers.main.app_factory import create_app, init_app
 from tassaron_flask.helpers.main.plugins import db
 from tassaron_flask.models.shop.inventory_models import *
 import tempfile
