+++ tests/test_sessions.py
@@ -2,7 +2,7 @@ import os
 import tempfile
 import pytest
 import json
-from tassaron_flask.models.main import create_app, init_app
+from tassaron_flask.helpers.main.app_factory import create_app, init_app
 from tassaron_flask.helpers.main.plugins import db
 from tassaron_flask.models.main.models import User
 from tassaron_flask.models.shop.inventory_models import *
